Refactor, move functions to better suited files, use temp dir for upload
