+++ server.go
@@ -37,11 +37,6 @@ func initHandlers(app *App, enableInternalApis bool) *echo.Echo {
 	e.GET("/languages", handleLanguages)
 	e.GET("/status", handleStatus)
 
-	e.POST("/learn", authUser(handleLearn))
-	e.POST("/learn/upload", authUser(handleLearnFileUpload))
-	e.POST("/train/:langCode", authUser(handleTrain))
-	e.POST("/train/bulk/:langCode", authUser(handleTrainBulk))
-
 	e.GET("/", handleIndex)
 
 	e.GET("/*", echo.WrapHandler(app.fs.FileServer()))
@@ -49,6 +44,11 @@ func initHandlers(app *App, enableInternalApis bool) *echo.Echo {
 	if enableInternalApis {
 		e.POST("/sync/download/:langCode/enable", handleEnableDownload)
 		e.POST("/sync/download/:langCode/disable", handleDisableDownload)
+
+		e.POST("/learn", authUser(handleLearn))
+		e.POST("/learn/upload/:langCode", authUser(handleLearnFileUpload))
+		e.POST("/train/:langCode", authUser(handleTrain))
+		e.POST("/train/bulk/:langCode", authUser(handleTrainBulk))
 	}
 
 	e.Use(middleware.Recover())
